Taxonomy

Documentation for Taxonomy.

Index

Taxonomy.DOI

Taxonomy.NoDOI

Taxonomy.UnusualDOI

Taxonomy.UsualDOI

Taxonomy.valid_doi

DOI

Taxonomy.DOI — Type

Alias for UsualDOI.

Taxonomy.UsualDOI — Type

Construct a validated DOI

Most valid DOIs (not all) can be simply validated via a regular expression.

Arguments

doi::String: a DOI without resolver (e.g. without doi.org), capitalization does not matter

fallback::String: an optional fallback link where one maybe can find the content in case the doi fails

julia> DOI("10.5281/zenodo.6719627")
UsualDOI{String, Missing}("10.5281/ZENODO.6719627", missing)

julia> DOI("10.5281/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UsualDOI{String, String}("10.5281/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.UnusualDOI — Type

Construct an unvalidated DOI

You should prefer an validated UsualDOI but if you have tested the DOI and are sure it links were it supposed to link, go ahead and create an unvalidated doi.

julia> UnusualDOI("weird10.5281doi/zenodo.6719627")
UnusualDOI{String, Missing}("WEIRD10.5281DOI/ZENODO.6719627", missing)

julia> UnusualDOI("weird10.5281doi/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UnusualDOI{String, String}("WEIRD10.5281DOI/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.NoDOI — Type

What to do if there is no doi

Last resort if there is no DOI. Than we save other metadata, similar to BibTex.

Arguments

url::String: an link where one maybe can find the content in case the doi fails

author::String: like in BibTex, e.g. "Peikert, Aaron and Ernst, Maximilian S. and Bode, Clifford"

date::Union{Date, Missing}: optional date

year::Union{Int64}: optional if date is supplied

journal::String

other::Dict: more BibTexlike metadata

NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    date = Date("2022-06-24"), # year is inferred
    journal = "No Real Journal",
    awesome = "Yes", # other metadata
    software = "naturally", # some more metadata
    citations = 500
)
NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    year = 2022, # date is omitted
    journal = "No Real Journal"
)

Taxonomy.valid_doi — Function

Validate DOI via Regex

Regular expression taken from:

https://www.crossref.org/blog/dois-and-matching-regular-expressions/