API Reference

Public

Taxonomy.All — Type

All()

Select the full Lineage.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> lineage[All()] === lineage
true

Taxonomy.Between — Type

Between(first, last)

Select lineage items from first through last, where each boundary can be a position or canonical rank. String rank names are converted to Symbols.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> taxid.(lineage[Between(:order, :family)])
6-element Vector{Int64}:
   9443
 376913
 314293
   9526
 314295
   9604

Taxonomy.Cols — Type

Cols(cols...)

Select lineage items by multiple positions or canonical ranks. Items must be requested from higher to lower rank.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> taxid.(lineage[Cols(:class, :species)])
2-element Vector{Int64}:
 40674
  9606

Taxonomy.From — Type

From(first)

Select lineage items from first through the end, where first can be a position or canonical rank. String rank names are converted to Symbols.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> first(taxid.(lineage[From(:family)])), last(taxid.(lineage[From(:family)]))
(9604, 9606)

Taxonomy.Lineage — Type

Lineage{T<:AbstractTaxon} <: AbstractVector{T}

A type that stores lineage information in Vector-like format. T represents element types, Taxon or UnclassifiedTaxon.

getindex is overloaded to get Taxon values. Symbols such as :domain, :superkingdom, :realm, :family, :genus, :species in CanonicalRanks can be used. Also, Between, From, Until, Cols and All selectors can be used in more complex rank selection scenarios.

Once reformatted, it cannot be reformatted again. The status can be checked using isreformatted(lineage).

Examples

julia> lineage = Lineage(Taxon(9606));

julia> lineage[:species]
9606 [species] Homo sapiens

julia> lineage[:class]
40674 [class] Mammalia

Taxonomy.Rank — Type

Rank(sym::Symbol)

Return CanonicalRank(sym) if sym is in CanonicalRanks. Return UnCanonicalRank(sym) if not. CanonicalRank(sym) can be used for isless comparison. :domain, :superkingdom, and :realm are treated as the same top-rank level.

Taxonomy.Rank — Method

Rank(taxon::Taxon)

Return CanonicalRank made from rank(taxon) if rank(taxon) is in CanonicalRanks. Return UnCanonicalRank(rank) if not. CanonicalRank(taxon) can be used for isless comparison. :domain, :superkingdom, and :realm can all be handled as canonical aliases for the same top-rank level.

Taxonomy.Taxon — Type

Taxon(taxid::Int, db::Taxonomy.DB)
Taxon(taxid::Int)

Construct a Taxon from its taxid. Omitting db automatically calls current_db(), which is usually the database that was last created.

Examples

julia> Taxon(9606)
9606 [species] Homo sapiens

julia> Taxon(9606, db)
9606 [species] Homo sapiens

Taxonomy.Until — Type

Until(last)

Select lineage items from the beginning through last, where last can be a position or canonical rank. String rank names are converted to Symbols.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> last(taxid.(lineage[Until(:kingdom)]))
33208

AbstractTrees.children — Method

children(taxon::Taxon)

Return the vector of Taxon objects that are children of the given Taxon object.

Examples

julia> sort(taxid.(children(Taxon(9606)))) == [63221, 741158]
true

AbstractTrees.isdescendant — Method

isdescendant(descendant::Taxon, ancestor::Taxon)

Return true if the former taxon is a descendant of the latter taxon.

Examples

julia> isdescendant(Taxon(9606), Taxon(9605))
true

julia> isdescendant(Taxon(9606), Taxon(10239))
false

AbstractTrees.parent — Method

AbstractTrees.parent(taxon::Taxon)

Return the Taxon object that is the parent of the given Taxon object.

Examples

julia> AbstractTrees.parent(Taxon(9606))
9605 [genus] Homo

Base.get — Method

get(lineage::Lineage, idx::Union{Int,Symbol}, default)

Return the Taxon object stored at the given position or rank (i.e. :phylum), or the given default value if no matching item is present.

Examples

julia> lineage = Lineage(Taxon(9606));

julia> get(lineage, :class, nothing)
40674 [class] Mammalia

julia> get(lineage, :strain, nothing) === nothing
true

Base.get — Method

get(db::Taxonomy.DB, taxid::Int, default)

Return the Taxon object stored for the given taxid, or the given default value if no mapping for the taxid is present.

Examples

julia> get(db, 9606, nothing)
9606 [species] Homo sapiens

julia> get(db, 99999999, nothing) === nothing
true

Taxonomy.current_db! — Method

current_db!(db::Taxonomy.DB)

Set db as the current active database.

Taxonomy.current_db — Method

current_db()

Return the current active database or the last database that got created.

Taxonomy.isancestor — Method

isancestor(ancestor::Taxon, descendant::Taxon)

Return true if the former taxon is an ancestor of the latter taxon.

Examples

julia> isancestor(Taxon(9605), Taxon(9606))
true

julia> isancestor(Taxon(10239), Taxon(9606))
false

Taxonomy.isreformatted — Method

isreformatted(lineage::Lineage)

Return true if lineage is already reformatted.

Examples

julia> isreformatted(Lineage(Taxon(9606)))
false

julia> ranks = [:domain, :phylum, :class, :order, :family, :genus, :species];

julia> isreformatted(reformat(Lineage(Taxon(9606)), ranks))
true

Taxonomy.lca — Method

lca(taxa::Vector{Taxon})
lca(taxa::Taxon...)

Return the Taxon object that is the lowest common ancestor of the given set of Taxons. At least one Taxon is required.

Examples

julia> human = Taxon(9606); gorilla = Taxon(9593);

julia> lca(human, gorilla)
207598 [subfamily] Homininae

julia> lca([human, gorilla])
207598 [subfamily] Homininae

julia> lca(human)
9606 [species] Homo sapiens

Taxonomy.name — Method

name(taxon::AbstractTaxon)

Return the name of the given Taxon object. It also works for an UnclassifiedTaxon object.

Examples

julia> name(Taxon(9606))
"Homo sapiens"

julia> name(UnclassifiedTaxon(:subspecies, Taxon(9606)))
"unclassified Homo sapiens subspecies"

Taxonomy.name2taxids — Method

name2taxids(name::AbstractString, db::Taxonomy.DB)
name2taxids(name::AbstractString)

Return a Vector of taxid from its name. name must match to the scientific name exactly. If multiple hits are found, return a multi-element Vector. If not, 1- or 0-element Vector. Omitting db automatically calls current_db(), which is usually the database that was last created.

Examples

julia> name2taxids("Homo")
1-element Vector{Int64}:
 9605

julia> name2taxids("ThisNameDoesNotExist")
Int64[]

Taxonomy.namedtuple — Method

namedtuple(lineage::Lineage; kwargs...)

Return a NamedTuple whose filednames is ranks (in the CanonicalRanks) of the lineage. This function is useful for converting Lineage to DataFrame, for example.

Arguments

fill_by_missing::Bool = false - If true, fills missing instead of UnclassifiedTaxon.

Taxonomy.print_lineage — Method

print_lineage(lineage::Lineage; kwargs...)
print_lineage(io::IO, lineage::Lineage; kwargs...)

Print a formatted representation of the lineage to the given IO object.

Arguments

delim::AbstractString = ";" - The delimiter between taxon fields.

fill::Bool = false - If true, prints UnclassifiedTaxon. only availavle when skip is false.

skip::Bool = false - If true, skip printing UnclassifiedTaxon and delimiter.

Taxonomy.rank — Method

rank(taxon::AbstractTaxon)

Return the rank of the given Taxon object. It also works for an UnclassifiedTaxon object.

Examples

julia> rank(Taxon(9606))
:species

julia> rank(UnclassifiedTaxon(:subspecies, Taxon(9606)))
:subspecies

Taxonomy.reformat — Method

reformat(l::Lineage, ranks::Vector{Symbol})

Return the Lineage object reformatted according to the given canonical ranks. Non-canonical ranks and multiple aliases for the same rank slot are not supported. If there is no corresponding taxon in the lineage to the rank, UnclassifiedTaxon will be stored. Once a Lineage is reformatted, it cannot be reformatted again.

Taxonomy.similarnames — Method

similarnames(query::AbstractString, db::Taxonomy.DB; distance::StringDistances.StringDistance=StringDistances.Levenshtein(), threshold::Float64=0.8)
similarnames(query::AbstractString; distance::StringDistances.StringDistance=StringDistances.Levenshtein(), threshold::Float64=0.8)

Find names similar to the query and return a Vector of NamedTuples with taxid, name, and similarity. The similarities are calculated by the Lavenshtein distance by default. It can also be changed to other distances defined in StringDistaces package by specifying it in the distance argument. Omitting db automatically calls current_db(), which is usually the database that was last created.

Taxonomy.source — Method

source(taxon::UnclassifiedTaxon)

Return the source Taxon used to construct an UnclassifiedTaxon.

Examples

julia> source(UnclassifiedTaxon(:subspecies, Taxon(9606)))
9606 [species] Homo sapiens

Taxonomy.taxid — Method

taxid(taxon::Taxon)

Return the taxid of the given Taxon object.

Examples

julia> taxid(Taxon(9606))
9606

Internal

Taxonomy.DB — Type

Taxonomy.DB

Constructors

DB(nodes_dmp::String, names_dmp::String)

Create DB(taxonomy database) object from nodes.dmp and names.dmp files.

Base.isless — Method

isless(taxon::AbstractTaxon, rank::CanonicalRank)

Example

julia> Taxon(9606 , db) < Rank(:genus)
true

Return true if the rank of the former Taxon is less than the later rank.