Taxonomy

Documentation for Taxonomy.

Index

Taxonomy.CFA

Taxonomy.DOI

Taxonomy.ExtensiveMeta

Taxonomy.IncompleteMeta

Taxonomy.J

Taxonomy.Judgement

Taxonomy.MinimalMeta

Taxonomy.NoDOI

Taxonomy.NoLocation

Taxonomy.UnusualDOI

Taxonomy.UsualDOI

Taxonomy.MetaData

Taxonomy.NoJudgement

Taxonomy.apa

Taxonomy.author

Taxonomy.certainty

Taxonomy.generate_id

Taxonomy.journal

Taxonomy.json

Taxonomy.location

Taxonomy.rating

Taxonomy.url

Taxonomy.valid_doi

Taxonomy.year

Taxon

CFA

Missing docstring.

Missing docstring for Factor. Check Documenter's build log for details.

Taxonomy.CFA — Type

CFA Taxon. Consists of Factors (measurement model) and a graph from StenoGraphs (structural model).

Arguments

measurement_model: Vector of Factors.

structural_model: Graph from StenoGraphs package. Defines the latent relations between the factors of measurement_model.

using StenoGraphs
using Taxonomy

factor1 = Factor(nobserved = 2, loadings = [1, 0.4])
factor2 = Factor(nobserved = 2, loadings = [0.7, 0.3])

graph = @StenoGraph begin
    # latent regressions
    fac1 → fac2
end

CFA(measurement_model = [factor1, factor2], 
structural_model = graph )

# output
CFA(Judgement{Vector{Factor}}(Factor[Factor(Judgement{Int64}(2, 1.0, missing), Judgement{Vector{Float64}}([1.0, 0.4], 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing)), Factor(Judgement{Int64}(2, 1.0, missing), Judgement{Vector{Float64}}([0.7, 0.3], 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing), Judgement{Int64}(0, 1.0, missing))], 1.0, missing), Judgement{Vector{DirectedEdge{SimpleNode{Symbol}, SimpleNode{Symbol}}}}(fac1 → fac2
, 1.0, missing))

ID

Taxonomy.generate_id — Function

Generate an Entry ID

To create links between entries we need a stable reference point. This ID is generated initially from url(location) and if the url is missing, it is generated randomly. After the ID is generated once, it is saved with the Record and should not be changed.

Judgement

Taxonomy.Judgement — Type

A judgment about any parameter etc.

Arguments

rating: The rating, e.g. "CFA" or 1.0.

certainty: If uncertain, a number between 0.0 and 1.0 (0-100%)

location: optional, Where in the Paper PDF was the location retieved, e.g. section, page, table number, figure number.

julia> Judgement(1.0, .99, "Figure 1");

Taxonomy.J — Type

Shorthand for Judgement

Taxonomy.NoJudgement — Function

Abstaining from any judgement.

This implies that your best guess is missing and you are absolutely uncertain about this judgement.

julia> NoJudgement()
Judgement{Missing}(missing, 0.0, missing)

Taxonomy.rating — Function

Extract rating from Judgement.

If rating is called on a Judgement it returns the rating, on everything it returns identity.

Taxonomy.location — Method

Extract location from Judgement.

Taxonomy.certainty — Function

Extract certainty from Judgement.

Metadata

Taxonomy.MetaData — Function

Save metadata.

Can be from complete minimal metadata, incomplete metadata or preferably from DOI.

julia> min = MetaData("Peikert, Aaron", 2022, "Journal of Statistical Software");

julia> incomplete = MetaData("Peikert, Aaron", 2022, missing);

julia> extensive = MetaData(DOI("10.5281/zenodo.6719627"));

Taxonomy.apa — Function

Get an APA citation.

julia> apa(DOI("10.5281/zenodo.6719627"))
"Ernst, M. S., &amp; Peikert, A. (2022). <i>StructuralEquationModels.jl</i> (Version v0.1.0) [Computer software]. Zenodo. https://doi.org/10.5281/ZENODO.6719627"

Taxonomy.json — Function

Get a Citeproc JSON.

CSL JSON Documentation

CSL JSON can be read by Zotero and automatically generated by doi.org from DOI. All availible information are included and saved in a Dict.

julia> json(DOI("10.5281/zenodo.6719627"))
Dict{String, Any} with 11 entries:
  "publisher" => "Zenodo"
  "issued"    => Dict{String, Any}("date-parts"=>Any[Any[2022, 6, 24]])
  "author"    => Any[Dict{String, Any}("family"=>"Ernst", "given"=>"Maximilian …
  "id"        => "https://doi.org/10.5281/zenodo.6719627"
  "copyright" => "MIT License"
  "version"   => "v0.1.0"
  "DOI"       => "10.5281/ZENODO.6719627"
  "URL"       => "https://zenodo.org/record/6719627"
  "title"     => "StructuralEquationModels.jl"
  "abstract"  => "StructuralEquationModels v0.1.0 This is a package for Structu…
  "type"      => "book"

Taxonomy.author — Function

Extract the author.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> author(doi)
"Frank, Henry S."

Taxonomy.year — Function

Extract the year.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> year(doi)
1970

Taxonomy.journal — Function

Extract the journal.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> journal(doi)
"Science"

Taxonomy.MinimalMeta — Type

A representation of the most important metadata.

julia> min = MetaData("Peikert, Aaron", 2022, "Journal of Statistical Software");

julia> typeof(min)
MinimalMeta

Taxonomy.IncompleteMeta — Type

A representation of Metadata when we can not even capture the most important metadata.

julia> incomplete = MetaData(missing, 2022, "Journal of Statistical Software");

julia> typeof(incomplete)
IncompleteMeta

Taxonomy.ExtensiveMeta — Type

The metadata we can gather from doi.org.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> typeof(doi)
ExtensiveMeta{MinimalMeta}

DOI

Taxonomy.DOI — Type

Alias for UsualDOI.

Taxonomy.UsualDOI — Type

Construct a validated DOI

Most valid DOIs (not all) can be simply validated via a regular expression.

Arguments

doi::String: a DOI without resolver (e.g. without doi.org), capitalization does not matter

fallback::String: an optional fallback link where one maybe can find the content in case the doi fails

julia> DOI("10.5281/zenodo.6719627")
UsualDOI{String, Missing}("10.5281/ZENODO.6719627", missing)

julia> DOI("10.5281/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UsualDOI{String, String}("10.5281/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.UnusualDOI — Type

Construct an unvalidated DOI

You should prefer an validated UsualDOI but if you have tested the DOI and are sure it links were it supposed to link, go ahead and create an unvalidated doi.

julia> UnusualDOI("weird10.5281doi/zenodo.6719627")
UnusualDOI{String, Missing}("WEIRD10.5281DOI/ZENODO.6719627", missing)

julia> UnusualDOI("weird10.5281doi/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UnusualDOI{String, String}("WEIRD10.5281DOI/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.NoDOI — Type

What to do if there is no doi

Last resort if there is no DOI. Than we save other metadata, similar to BibTex.

Arguments

url::String: an link where one maybe can find the content in case the doi fails

author::String: like in BibTex, e.g. "Peikert, Aaron and Ernst, Maximilian S. and Bode, Clifford"

date::Union{Date, Missing}: optional date

year::Union{Int64}: optional if date is supplied

journal::String: The outlet of the publication

other::Dict: more BibTexlike metadata

NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    date = Date("2022-06-24"), # year is inferred
    journal = "No Real Journal",
    awesome = "Yes", # other metadata
    software = "naturally", # some more metadata
    citations = 500
)
NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    year = 2022, # date is omitted
    journal = "No Real Journal"
)

Taxonomy.NoLocation — Type

When everything fails.

This is a placeholder if really no location can be found.

Taxonomy.url — Function

Get URL from location.

julia> url(DOI("10.1126/SCIENCE.169.3946.635"))
"https://doi.org/10.1126/SCIENCE.169.3946.635"

location = NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    year = 2022, # date is omitted
    journal = "No Real Journal"
)

url(location)

# output

"https://github.com/StructuralEquationModels/StructuralEquationModels.jl"

Taxonomy.valid_doi — Function

Validate DOI via Regex

Regular expression taken from:

https://www.crossref.org/blog/dois-and-matching-regular-expressions/