Taxonomy

Documentation for Taxonomy.

Index

Taxonomy.DOI

Taxonomy.ExtensiveMeta

Taxonomy.IncompleteMeta

Taxonomy.MinimalMeta

Taxonomy.NoDOI

Taxonomy.UnusualDOI

Taxonomy.UsualDOI

Taxonomy.MetaData

Taxonomy.apa

Taxonomy.author

Taxonomy.journal

Taxonomy.json

Taxonomy.url

Taxonomy.valid_doi

Taxonomy.year

Metadata

Taxonomy.MetaData — Function

Save metadata.

Can be from complete minimal metadata, incomplete metadata or preferably from DOI.

julia> min = MetaData("Peikert, Aaron", 2022, "Journal of Statistical Software");

julia> incomplete = MetaData("Peikert, Aaron", 2022, missing);

julia> extensive = MetaData(DOI("10.5281/zenodo.6719627"));

Taxonomy.apa — Function

Get an APA citation.

julia> apa(DOI("10.5281/zenodo.6719627"))
"Ernst, M. S., &amp; Peikert, A. (2022). <i>StructuralEquationModels.jl</i> (Version v0.1.0) [Computer software]. Zenodo. https://doi.org/10.5281/ZENODO.6719627"

Taxonomy.json — Function

Get a Citeproc JSON.

CSL JSON Documentation

CSL JSON can be read by Zotero and automatically generated by doi.org from DOI. All availible information are included and saved in a Dict.

julia> json(DOI("10.5281/zenodo.6719627"))
Dict{String, Any} with 11 entries:
  "publisher" => "Zenodo"
  "issued"    => Dict{String, Any}("date-parts"=>Any[Any[2022, 6, 24]])
  "author"    => Any[Dict{String, Any}("family"=>"Ernst", "given"=>"Maximilian …
  "id"        => "https://doi.org/10.5281/zenodo.6719627"
  "copyright" => "MIT License"
  "version"   => "v0.1.0"
  "DOI"       => "10.5281/ZENODO.6719627"
  "URL"       => "https://zenodo.org/record/6719627"
  "title"     => "StructuralEquationModels.jl"
  "abstract"  => "StructuralEquationModels v0.1.0 This is a package for Structu…
  "type"      => "book"

Taxonomy.author — Function

Extract the author.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> author(doi)
"Frank, Henry S."

Taxonomy.year — Function

Extract the year.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> year(doi)
1970

Taxonomy.journal — Function

Extract the journal.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> journal(doi)
"Science"

Taxonomy.MinimalMeta — Type

A representation of the most important metadata.

julia> min = MetaData("Peikert, Aaron", 2022, "Journal of Statistical Software");

julia> typeof(min)
MinimalMeta

Taxonomy.IncompleteMeta — Type

A representation of Metadata when we can not even capture the most important metadata.

julia> incomplete = MetaData(missing, 2022, "Journal of Statistical Software");

julia> typeof(incomplete)
IncompleteMeta

Taxonomy.ExtensiveMeta — Type

The metadata we can gather from doi.org.

julia> doi = MetaData(DOI("10.1126/SCIENCE.169.3946.635"));

julia> typeof(doi)
ExtensiveMeta

DOI

Taxonomy.DOI — Type

Alias for UsualDOI.

Taxonomy.UsualDOI — Type

Construct a validated DOI

Most valid DOIs (not all) can be simply validated via a regular expression.

Arguments

doi::String: a DOI without resolver (e.g. without doi.org), capitalization does not matter

fallback::String: an optional fallback link where one maybe can find the content in case the doi fails

julia> DOI("10.5281/zenodo.6719627")
UsualDOI{String, Missing}("10.5281/ZENODO.6719627", missing)

julia> DOI("10.5281/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UsualDOI{String, String}("10.5281/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.UnusualDOI — Type

Construct an unvalidated DOI

You should prefer an validated UsualDOI but if you have tested the DOI and are sure it links were it supposed to link, go ahead and create an unvalidated doi.

julia> UnusualDOI("weird10.5281doi/zenodo.6719627")
UnusualDOI{String, Missing}("WEIRD10.5281DOI/ZENODO.6719627", missing)

julia> UnusualDOI("weird10.5281doi/zenodo.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")
UnusualDOI{String, String}("WEIRD10.5281DOI/ZENODO.6719627", "https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Taxonomy.NoDOI — Type

What to do if there is no doi

Last resort if there is no DOI. Than we save other metadata, similar to BibTex.

Arguments

url::String: an link where one maybe can find the content in case the doi fails

author::String: like in BibTex, e.g. "Peikert, Aaron and Ernst, Maximilian S. and Bode, Clifford"

date::Union{Date, Missing}: optional date

year::Union{Int64}: optional if date is supplied

journal::String: The outlet of the publication

other::Dict: more BibTexlike metadata

NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    date = Date("2022-06-24"), # year is inferred
    journal = "No Real Journal",
    awesome = "Yes", # other metadata
    software = "naturally", # some more metadata
    citations = 500
)
NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    year = 2022, # date is omitted
    journal = "No Real Journal"
)

Taxonomy.url — Function

Get URL from location.

julia> url(DOI("10.1126/SCIENCE.169.3946.635"))
"https://doi.org/10.1126/SCIENCE.169.3946.635"

location = NoDOI(
    url = "https://github.com/StructuralEquationModels/StructuralEquationModels.jl",
    author = "Ernst, Maximilian Stefan and Peikert, Aaron",
    title = "StructuralEquationModels.jl: A fast and flexible SEM framework",
    year = 2022, # date is omitted
    journal = "No Real Journal"
)

url(location)

# output

"https://github.com/StructuralEquationModels/StructuralEquationModels.jl"

Taxonomy.valid_doi — Function

Validate DOI via Regex

Regular expression taken from:

https://www.crossref.org/blog/dois-and-matching-regular-expressions/