API Reference

Public

AbstractTrees.children — Method

children(taxon::Taxon)

Return the vector of Taxon objects that are children of the given Taxon object.

AbstractTrees.isdescendant — Method

isdescendant(descendant::Taxon, ancestor::Taxon)

Return true if the former taxon is a descendant of the latter taxon. This function is overloaded because native AbstractTrees.isdescendant is too slow

AbstractTrees.parent — Method

parent(taxon::Taxon)

Return the Taxon object that is the parent of the given Taxon object.

Taxonomy.isancestor — Method

isancestor(ancestor::Taxon, descendant::Taxon)

Return true if the former taxon is an ancestor of the latter taxon.

Taxonomy.lca — Method

lca(taxa::Vector{Taxon})
lca(taxa::Taxon...)

Return the Taxon object that is the lowest common ancestor of the given set of Taxons

Taxonomy.name — Method

name(taxon::AbstractTaxon)

Return the name of the given Taxon object. It also works for an UnclassifiedTaxon object.

Taxonomy.print_lineage — Method

print_lineage(lineage::Lineage; kwargs...)
print_lineage(io::IO, lineage::Lineage; kwargs...)

Print a formatted representation of the lineage to the given IO object.

Arguments

delim::AbstractString = ";" - The delimiter between taxon fields.

fill::Bool = false - If true, prints UnclassifiedTaxon. only availavle when skip is false

skip::Bool= false- If true, skip printingUnclassifiedTaxon` and delimiter.

Taxonomy.rank — Method

rank(taxon::AbstractTaxon)

Return the rank of the given Taxon object. It also works for an UnclassifiedTaxon object.

Taxonomy.reformat — Method

reformat(l::Lineage, ranks::Vector{Symbol})

Return the Lineage object reformatted according to the given ranks.

Taxonomy.taxid — Method

taxid(taxon::Taxon)

Return the taxid of the given Taxon object.

Internal

Taxonomy.DB — Method

Taxonomy.DB

Constructors

DB(nodes_dmp::String, names_dmp::String)
DB(db_path::String, nodes_dmp::String, names_dmp::String)

Create DB(taxonomy database) object from nodes.dmp and names.dmp files. You can specify the paths of the nodes.dmp and names.dmp files, or the directory where they exist and the names.

Base.get — Method

get(db::Taxonomy.DB, idx::Union{Int,Symbol}, default)

Return the Taxon object stored for the given taxid or rank (i.e. :phylum), or the given default value if no mapping for the taxid is present.

Base.get — Method

get(db::Taxonomy.DB, taxid::Int, default)

Return the Taxon object stored for the given taxid, or the given default value if no mapping for the taxid is present.

Base.get — Method

get(db::Taxonomy.DB, name::String, default)

Return the Taxon object stored for the given name, or the given default value if no mapping for the name is present.